+++ api/v1alpha2/helpers.go
@@ -1,0 +1,8 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
+package v1alpha2
+
+func PointerOf[A any](a A) *A {
+	return &a
+}