+++ controllers/agentpool_controller_tokens.go
@@ -105,8 +105,8 @@ func (r *AgentPoolReconciler) createAgentPoolTokens(ctx context.Context, ap *age
 		ap.instance.Status.AgentTokens = append(ap.instance.Status.AgentTokens, &appv1alpha2.AgentToken{
 			Name:       at.Description,
 			ID:         at.ID,
-			CreatedAt:  at.CreatedAt.Unix(),
-			LastUsedAt: at.LastUsedAt.Unix(),
+			CreatedAt:  appv1alpha2.PointerOf(at.CreatedAt.Unix()),
+			LastUsedAt: appv1alpha2.PointerOf(at.LastUsedAt.Unix()),
 		})
 		ap.log.Info("Reconcile Agent Tokens", "msg", fmt.Sprintf("successfully created a new agent token %q %q", t, at.ID))
 		// UPDATE SECRET